Move SchemaRegistry into arcs.core.data package
Also makes its API slightly nicer (return non-nullable).

I'll need to call SchemaRegistry from the data package in a future PR, so I need to move it to avoid circular deps.
