+++ java/arcs/android/storage/database/DatabaseImpl.kt
@@ -42,10 +42,10 @@ import arcs.core.data.LARGEST_PRIMITIVE_TYPE_ID
 import arcs.core.data.PrimitiveType
 import arcs.core.data.RawEntity
 import arcs.core.data.Schema
+import arcs.core.data.SchemaRegistry
 import arcs.core.data.util.ReferencableList
 import arcs.core.data.util.ReferencablePrimitive
 import arcs.core.data.util.toReferencable
-import arcs.core.entity.SchemaRegistry
 import arcs.core.storage.Reference
 import arcs.core.storage.StorageKey
 import arcs.core.storage.StorageKeyParser
@@ -1709,9 +1709,7 @@ class DatabaseImpl(
     ): TypeId = when (fieldType) {
         is FieldType.Primitive -> fieldType.primitiveType.primitiveTypeId()
         is FieldType.EntityRef -> {
-            val schema = requireNotNull(SchemaRegistry.getSchema(fieldType.schemaHash)) {
-                "Unknown Schema with hash: ${fieldType.schemaHash} in SchemaRegistry"
-            }
+            val schema = SchemaRegistry.getSchema(fieldType.schemaHash)
             getSchemaTypeId(schema, database)
         }
         // TODO(b/156003617)
@@ -1719,9 +1717,7 @@ class DatabaseImpl(
             throw NotImplementedError("[FieldType.Tuple]s not currently supported.")
         is FieldType.ListOf -> getTypeId(fieldType.primitiveType, database)
         is FieldType.InlineEntity -> {
-            var schema = requireNotNull(SchemaRegistry.getSchema(fieldType.schemaHash)) {
-                "Unknown Schema with hash: ${fieldType.schemaHash} in SchemaRegistry"
-            }
+            val schema = SchemaRegistry.getSchema(fieldType.schemaHash)
             getSchemaTypeId(schema, database)
         }
     }
