+++ java/arcs/core/storage/api/DriverAndKeyConfigurator.kt
@@ -13,7 +13,7 @@ package arcs.core.storage.api
 
 import arcs.core.common.ArcId
 import arcs.core.data.CreatableStorageKey
-import arcs.core.entity.SchemaRegistry
+import arcs.core.data.SchemaRegistry
 import arcs.core.storage.CapabilitiesResolver
 import arcs.core.storage.DriverFactory
 import arcs.core.storage.StorageKeyParser
