+++ javatests/arcs/android/storage/ttl/PeriodicCleanupTaskTest.kt
@@ -13,7 +13,7 @@ import arcs.core.data.Capability.Ttl
 import arcs.core.entity.DummyEntity
 import arcs.core.entity.HandleSpec
 import arcs.core.entity.ReadWriteCollectionHandle
-import arcs.core.entity.SchemaRegistry
+import arcs.core.data.SchemaRegistry
 import arcs.core.entity.awaitReady
 import arcs.core.host.EntityHandleManager
 import arcs.core.storage.StoreWriteBack
