+++ javatests/arcs/android/systemhealth/testapp/TestApplication.kt
@@ -16,7 +16,7 @@ import androidx.work.Configuration
 import arcs.android.storage.database.AndroidSqliteDatabaseManager
 import arcs.android.util.connectMemoryStatsPipe
 import arcs.android.util.initLogForAndroid
-import arcs.core.entity.SchemaRegistry
+import arcs.core.data.SchemaRegistry
 import arcs.core.storage.api.DriverAndKeyConfigurator
 import arcs.core.storage.driver.RamDisk
