+++ src/tools/tests/goldens/generated-schemas.jvm.kt
@@ -9,10 +9,10 @@ package arcs.golden
 // Current implementation doesn't support optional field detection
 
 import arcs.core.data.*
+import arcs.core.data.SchemaRegistry
 import arcs.core.data.util.ReferencablePrimitive
 import arcs.core.data.util.toReferencable
 import arcs.core.entity.Reference
-import arcs.core.entity.SchemaRegistry
 import arcs.core.entity.Tuple1
 import arcs.core.entity.Tuple2
 import arcs.core.entity.Tuple3
